Refactor hash-consing classes to use immutable package structure
