+++ foresight/src/main/scala/foresight/eqsat/hashCons/immutable/SlottedUnionFind.scala
@@ -1,4 +1,4 @@
-package foresight.eqsat.hashCons
+package foresight.eqsat.hashCons.immutable
 
 import foresight.eqsat.{EClassCall, EClassRef}
